First pass at refactor
